Ajuste no Repository e Entity da Sis_UsuarioFilter, pois precisou incluir o campo ID e a anotação Entity.

+++ src/main/java/mierim/model/repository/sistema/Sis_UsuarioFilterRepository.java
@@ -3,5 +3,5 @@ package mierim.model.repository.sistema;
 import mierim.model.entity.sistema.Sis_UsuarioFilter;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Sis_UsuarioFilterRepository extends JpaRepository<Sis_UsuarioFilter, Long> {
+public interface Sis_UsuarioFilterRepository extends JpaRepository<Sis_UsuarioFilter, Integer> {
 }
